# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for France Open Data Portal explorer
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export type ChatMessageRole =
  (typeof ChatMessageRole)[keyof typeof ChatMessageRole];

export const ChatMessageRole = {
  user: "user",
  assistant: "assistant",
  system: "system",
} as const;

export interface ChatMessage {
  role: ChatMessageRole;
  content: string;
}

export interface ChatMessageRequest {
  messages: ChatMessage[];
  /**
   * Optional context about a dataset or search result
   * @nullable
   */
  context?: string | null;
}

export interface ChatMessageResponse {
  content: string;
  role: string;
  /**
   * AI reasoning trace (one loop)
   * @nullable
   */
  thinking?: string | null;
}